+++ lib/zenlish/wclasses/conjunctive_pronoun.rb
@@ -1,0 +1,12 @@
+require_relative 'pronoun'
+
+module Zenlish
+  module WClasses
+    # As a a relative pronoun, a conjunctive (connective) pronoun
+    # refers to some noun appearing before, but, in addition,
+    # it also joins two sentences together.
+    # Conjunctive pronouns are: who, what, whose, which.
+    class ConjunctivePronoun < Pronoun
+    end # class
+  end # module
+end # module