+++ spec/zenlish/support/minimal_lexicon.rb
@@ -21,7 +21,9 @@ preposition_than = $ZenlishLexicon.name2terminal['PrepositionThan']
 regular_verb = $ZenlishLexicon.name2terminal['RegularVerb']
 irregular_verb = $ZenlishLexicon.name2terminal['IrregularVerb']
 irregular_verb_be = $ZenlishLexicon.name2terminal['IrregularVerbBe']
+irregular_verb_say = $ZenlishLexicon.name2terminal['IrregularVerbSay']
 indefinite_pronoun = $ZenlishLexicon.name2terminal['IndefinitePronoun']
+conjunctive_pronoun = $ZenlishLexicon.name2terminal['ConjunctivePronoun']
 demonstrative_pronoun = $ZenlishLexicon.name2terminal['DemonstrativePronoun']
 demonstrative_determiner = $ZenlishLexicon.name2terminal['DemonstrativeDeterminer']
 definite_article = $ZenlishLexicon.name2terminal['DefiniteArticle']
@@ -29,8 +31,11 @@ cardinal = $ZenlishLexicon.name2terminal['Cardinal']
 comparative_particle = $ZenlishLexicon.name2terminal['ComparativeParticle']
 indefinite_quantifier = $ZenlishLexicon.name2terminal['IndefiniteQuantifier']
 subordinating_conjunction = $ZenlishLexicon.name2terminal['SubordinatingConjunction']
+
+colon = $ZenlishLexicon.name2terminal['Colon']
 comma = $ZenlishLexicon.name2terminal['Comma']
 dot = $ZenlishLexicon.name2terminal['Period']
+quote = $ZenlishLexicon.name2terminal['Quote']
 
 def add_entry(aLemma, aWordClass)
   entry = Zenlish::Lex::LexicalEntry.new(aLemma)
@@ -40,6 +45,7 @@ def add_entry(aLemma, aWordClass)
 end
 
 # Our minimalistic lexicon
+add_entry('about', preposition)
 add_entry('above', preposition)
 add_entry('alive', adjective)
 add_entry('all', indefinite_quantifier)
@@ -49,8 +55,10 @@ add_entry('be', auxiliary_be)
 add_entry('be', irregular_verb_be)
 add_entry('big', adjective)
 add_entry('do', auxiliary_do)
+add_entry('false', adjective)
 add_entry('far', adverb)
 add_entry('from', preposition)
+add_entry('hear', irregular_verb)
 add_entry('if', subordinating_conjunction)
 add_entry('in', preposition)
 add_entry('inside', preposition)
@@ -69,6 +77,7 @@ add_entry('people', common_noun)
 add_entry('person', common_noun)
 add_entry('place', common_noun)
 add_entry('same', adjective)
+add_entry('say', irregular_verb_say)
 add_entry('see', irregular_verb)
 add_entry('side', common_noun)
 add_entry('small', adjective)
@@ -82,10 +91,16 @@ add_entry('thing', common_noun)
 add_entry('this', demonstrative_determiner)
 add_entry('this', demonstrative_pronoun)
 add_entry('this one', demonstrative_pronoun)
+add_entry('to', preposition)
 add_entry('Tony', proper_noun)
 add_entry('touch', regular_verb)
+add_entry('true', adjective)
 add_entry('two', cardinal)
 add_entry('very', degree_adverb)
+add_entry('what',conjunctive_pronoun)
+add_entry('word', common_noun)
 
+add_entry(':', colon)
 add_entry(',', comma)
 add_entry('.', dot)
+add_entry('"', quote)